fix(admin): admin defaults not working
